fixed: update networkmodel to network types and move to shared types folder
