+++ src/background/service/userWallet.ts
@@ -12,17 +12,16 @@ import { keyringService, mixpanelTrack, openapiService } from 'background/servic
 import { createPersistStore } from 'background/utils';
 import { getStoragedAccount } from 'background/utils/getStoragedAccount';
 
-import { fclConfig } from '../fclConfig';
-import { findAddressWithSeed, findAddressWithPK } from '../utils/modules/findAddressWithPK';
-import { storage } from '../webapi';
-
 import type {
   WalletResponse,
   BlockchainResponse,
   ChildAccount,
   DeviceInfoRequest,
   FlowNetwork,
-} from './networkModel';
+} from '../../shared/types/network-types';
+import { fclConfig } from '../fclConfig';
+import { findAddressWithSeed, findAddressWithPK } from '../utils/modules/findAddressWithPK';
+import { storage } from '../webapi';
 
 interface UserWalletStore {
   wallets: Record<string, WalletResponse[]>;
