+++ src/shared/types/network-types.ts
