+++ src/shared/utils/type.ts
