+++ src/ui/views/Send/SendAmount.tsx
@@ -3,12 +3,12 @@ import { Box, Button, Typography, IconButton, CardMedia } from '@mui/material';
 import React, { useState, useEffect, useCallback } from 'react';
 import { useHistory, useLocation } from 'react-router-dom';
 
+import { type Contact } from '@/shared/types/network-types';
 import { type ActiveChildType } from '@/shared/types/wallet-types';
 import { withPrefix } from '@/shared/utils/address';
 import { LLHeader } from '@/ui/FRWComponent';
 import SlideRelative from '@/ui/FRWComponent/SlideRelative';
 import { type CoinItem } from 'background/service/coinList';
-import { type Contact } from 'background/service/networkModel';
 import { LLContactCard } from 'ui/FRWComponent';
 import { useWallet } from 'ui/utils';
