+++ src/ui/views/Swap/index.tsx
@@ -3,10 +3,10 @@ import React, { useState, useEffect, useCallback } from 'react';
 
 // import { useHistory, useLocation } from 'react-router-dom';
 // import ArrowBackIcon from '@mui/icons-material/ArrowBack';
+import { type Contact } from '@/shared/types/network-types';
 import { withPrefix } from '@/shared/utils/address';
 import { LLHeader } from '@/ui/FRWComponent';
 import { type CoinItem } from 'background/service/coinList';
-import { type Contact } from 'background/service/networkModel';
 import { LLSpinner } from 'ui/FRWComponent';
 import { useWallet } from 'ui/utils';
